% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecology.R
\name{assign_areas}
\alias{assign_areas}
\title{Assign areas to geographical coordinates.}
\usage{
assign_areas(data, what, longitude = "lon", latitude = "lat")
}
\arguments{
\item{data}{A data frame including spatial coordinates}

\item{what}{the database with the geographic areas. 
Currently, only "longhurst" and "LME" is accepted.}

\item{longitude}{The name of the variable in \code{data} with longitud
information, by default "lat"}

\item{latitude}{The name of the variable in \code{data} with latitude
information, by default "lat"}
}
\value{

}
\description{
Assign areas to geographical coordinates.
}
