% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecology.R
\name{assign_region}
\alias{assign_region}
\title{Assign region to geographical coordinates.}
\usage{
assign_region(data, what, longitude = "lon", latitude = "lat")
}
\arguments{
\item{data}{A data frame including spatial coordinates}

\item{what}{the database with the geographic areas. 
Currently, only "longhurst", "LME", "FAO" and "GFMC" are accepted.}

\item{longitude}{The name of the variable in \code{data} with longitud
information, by default "lat"}

\item{latitude}{The name of the variable in \code{data} with latitude
information, by default "lat"}
}
\value{
A vector with the corresponding information (what) for the coordinates in data.
}
\description{
Assign region to geographical coordinates.
}
