% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributionData.R
\name{getDistributionRecords}
\alias{getDistributionRecords}
\title{Get distribution records from on-line databases}
\usage{
getDistributionRecords(
  sp,
  limit = NULL,
  source = "gbif",
  year = NULL,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{sp}{The scientific name of the species}

\item{limit}{The maximum number of records, NULL gets all records.}

\item{source}{The data source (currently only "gbif" and "obis" are supported.)}

\item{year}{Range of years to be included in the search.}

\item{verbose}{boolean, do you want to know what's happening?}

\item{...}{May be used later}
}
\value{

}
\description{
Get distribution records from on-line databases
}
\examples{
getDistributionRecords(sp="Manta birostris", limit=300)
}
