% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tolerance.R
\name{getTolerance}
\alias{getTolerance}
\title{Estimate Tolerance to an environmental factor}
\usage{
getTolerance(x, what, split = NULL, ...)
}
\arguments{
\item{x}{}

\item{what}{}

\item{split}{}

\item{...}{}
}
\value{
The tolerance levels: absolute minimum and maximum and optimum range.
}
\description{
Estimate Tolerance to an environmental factor
}
