% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{image.map}
\alias{image.map}
\title{Nice image plots for maps}
\usage{
\method{image}{map}(
  lon,
  lat,
  z,
  zlim = NULL,
  center = 0,
  legend = TRUE,
  hires = FALSE,
  add = FALSE,
  nlevel = 1000,
  horizontal = FALSE,
  legend.shrink = 0.9,
  legend.width = 1.2,
  slice = NULL,
  legend.mar = ifelse(horizontal, 3.1, 5.1),
  legend.lab = NULL,
  graphics.reset = FALSE,
  bigplot = NULL,
  smallplot = NULL,
  legend.only = FALSE,
  col = rev(rainbow(nlevel/10, start = 0/6, end = 4/6)),
  lab.breaks = NULL,
  axis.args = NULL,
  legend.args = NULL,
  axes = TRUE,
  midpoint = FALSE,
  border = TRUE,
  lwd = 1,
  land = TRUE,
  land.col = "black",
  labels = TRUE,
  sea.col = "white",
  boundaries.col = "grey",
  grid = FALSE,
  grid.col = "white",
  grid.lwd = 0.5,
  ...
)
}
\arguments{
\item{lon}{vector or matrix of longitude values, following the conventions
of \code{image}.}

\item{lat}{vector or matrix of latitude values, following the conventions
of \code{image}.}

\item{z}{matrix of values, following the conventions of \code{image}.}

\item{center}{}

\item{legend}{}

\item{hires}{}

\item{add}{}

\item{nlevel}{}

\item{horizontal}{}

\item{legend.shrink}{}

\item{legend.width}{}

\item{slice}{}

\item{legend.mar}{}

\item{legend.lab}{}

\item{graphics.reset}{}

\item{bigplot}{}

\item{smallplot}{}

\item{legend.only}{}

\item{col}{}

\item{lab.breaks}{}

\item{axis.args}{}

\item{legend.args}{}

\item{axes}{}

\item{midpoint}{}

\item{border}{}

\item{lwd}{}

\item{land}{}

\item{land.col}{}

\item{labels}{}

\item{sea.col}{}

\item{boundaries.col}{}

\item{grid}{}

\item{grid.col}{}

\item{...}{}
}
\description{
This function plots an image for a georeferenced matrix, 
assuming x and y axes are longitude and latitude.
}
\examples{
load(anchovy)
image.map(lon=anchovy$lon, lat=anchovy$lat, z=anchovy$z)
}
\author{
Ricardo Oliveros-Ramos, modified from xxx's \code{image.plot} from 
the fields package
}
