% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_grid.R
\name{is_ocean}
\alias{is_ocean}
\alias{is_land}
\title{Are the points in the ocean (or land)?}
\usage{
is_ocean(lon, lat, hires = FALSE)

is_land(lon, lat, hires = FALSE)
}
\arguments{
\item{lon}{Vector of longitudes.}

\item{lat}{Vector of latitudes.}

\item{hires}{Boolean. Use high resolution mask to compute ocean/land?}
}
\value{
A logical vector. For is_ocean, TRUE when the point is on water, FALSE for land or NA. Similiar for is_land.
}
\description{
Are the points in the ocean (or land)?
}
