\name{kali-package}
\alias{kali-package}
\alias{kali}
\docType{package}
\title{
Tools for running BIOMOD models using netCDF data
}
\description{
Tools for running BIOMOD models using netCDF data
}
\details{
\tabular{ll}{
Package: \tab kali\cr
Type: \tab Package\cr
Version: \tab 0.4\cr
Date: \tab 2013-11-04\cr
License: \tab GLP-2\cr
}
}
\author{
Ricardo Oliveros-Ramos

Maintainer: Ricardo Oliveros-Ramos <ricardo.oliveros@gmail.com>
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
}
\examples{
}
