% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_grid.R
\name{make_grid}
\alias{make_grid}
\title{Make a rectangular grid}
\usage{
make_grid(
  lon = lon,
  lat = lat,
  dx = dx,
  dy = dy,
  n = 1,
  thr = 0.8,
  hires = FALSE,
  mask = TRUE
)
}
\arguments{
\item{lon}{Longitude of the left and right corners of the grid.}

\item{lat}{Latitude of the upper and lower corners of the grid.}

\item{dx}{Resolution in the longitude axis (in degrees).}

\item{dy}{Resolution in the latitude axis (in degrees).}

\item{n}{Number of points within a square to test ocean/land mask. See details.}

\item{hires}{Use high resolution coast line? Default to FALSE, TRUE will take considerly longer time to compute.}
}
\value{
A list, including the grid and land/ocean mask.
}
\description{
Make a rectangular grid
}
\examples{
grid = make_grid(lon=c(2, 9), lat=c(40,45), dx=1/12, dy=1/12, n=1)
plot(grid)
}
