% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{map_details}
\alias{map_details}
\title{Add map details to a plot.}
\usage{
map_details(
  primeMeridian = "center",
  hires = FALSE,
  col = "darkolivegreen4",
  interior = FALSE,
  axes = FALSE,
  border = FALSE,
  boundaries.col = "black",
  grid = FALSE,
  grid.col = "white",
  grid.lwd = 0.5,
  cex.axis = 0.75,
  fill = TRUE,
  boundary = TRUE,
  water = NULL,
  land = TRUE,
  countries = FALSE,
  nx = NULL,
  ny = nx,
  labels = TRUE,
  ...
)
}
\arguments{
\item{primeMeridian}{}

\item{hires}{}

\item{col}{}

\item{interior}{}

\item{axes}{}

\item{border}{}

\item{boundaries.col}{}

\item{grid}{}

\item{grid.col}{}

\item{cex.axis}{}

\item{fill}{}

\item{boundary}{}

\item{water}{}

\item{countries}{}

\item{nx}{}

\item{ny}{}

\item{labels}{}

\item{...}{}
}
\value{
Nothing, used for side effect of adding map details to a plot.
}
\description{
Add map details to a plot.
}
