% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.map}
\alias{plot.map}
\title{Nice scatter plots for georeferenced data}
\usage{
\method{plot}{map}(x, y = NULL, xlim = NULL, ylim = NULL, domain = NA,
  center = 0, hires = FALSE, land.col = "darkolivegreen4",
  sea.col = "aliceblue", boundaries.col = "black", grid.col = "white",
  grid = TRUE, cex = 0.5, pch = 19, main = NULL, add = FALSE,
  axes = TRUE, border = !axes, asp = NA, axs = "i", xaxs = axs,
  yaxs = axs, cex.axis = 0.75, interior = FALSE, fill = TRUE, ...)
}
\arguments{
\item{x}{vector of longitude values, or a data.frame containing variables named 'lon'
and 'lat'}

\item{y}{vector of latitude values.}

\item{\dots}{Additional parameters to be passed to \code{points}.}
}
\description{
This function makes and scatter plot given the longitude and latitude
of the data points.
}
\examples{
load(mackerel)
image.map(mackerel)
image.map(domain="peru")
}
\author{
Ricardo Oliveros-Ramos
}
