% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extras.R
\name{sample.weighted}
\alias{sample.weighted}
\title{Weighted random sampling with a reservoir}
\usage{
sample.weighted(x, size, prob)
}
\arguments{
\item{x}{a vector of one or more elements from which to choose.}

\item{size}{a non-negative integer giving the number of items to choose.}

\item{prob}{a vector of weights for obtaining the elements of the vector being sampled.}
}
\value{
A vector of length \code{size} with elements drawn from \code{x}
}
\description{
Implementation of the Weighted random sampling with a reservoir (without replacement)
(Efraimidis & Spirakis, 2006) algorithm.
}
\examples{
N = 1000
x = seq_len(N)
prob = c(rep(0.1, N/2), rep(1, N/2))
x_sample = sample.weighted(x=x, prob=prob, size=N/2)
hist(x_sample)
}
\references{
Efraimidis & Spirakis (2006). Weighted random sampling with a reservoir
}
