% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_grid.R
\name{update_grid}
\alias{update_grid}
\title{Update grid}
\usage{
update_grid(grid, thr)
}
\arguments{
\item{grid}{An object of class 'grid'}

\item{thr}{A value between 0 and 1, used as threshold to define an ocean cell.}
}
\value{
A grid object.
}
\description{
Update grid
}
