% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecology.R
\name{add_region}
\alias{add_region}
\title{Add known areas to a map}
\usage{
add_region(
  what,
  region = NULL,
  subset = NULL,
  col = "red",
  list = FALSE,
  plot = TRUE,
  ...
)
}
\arguments{
\item{what}{the database with the geographic areas. 
Currently, only "longhurst", "LME", "FAO" and "GFMC" are accepted.}

\item{subset}{The polygon within the database especified in 'what'.}

\item{col}{The color of the polygons.}

\item{list}{Boolean, list the available regions in the dataset.}

\item{plot}{}

\item{...}{Additional arguments to the plot function (currently not used)}
}
\value{
Invisibly, the polygon to plot.
}
\description{
Add known areas to a map
}
\examples{
plot.map("world")
add_region(what="longhurst", col=1:4)
add_region(what="longhurst", region="ETRA", col="red")
}
