% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kali-main.R
\docType{package}
\name{kali-package}
\alias{kali-package}
\alias{kali}
\title{Tools for habitat and niche modeling.}
\description{
Tools for habitat and niche modeling.
}
\details{
\tabular{ll}{ Package: \tab kali\cr Type: \tab Package\cr Version: \tab
0.1\cr Date: \tab 2015-08-09\cr License: \tab GPL-2\cr }
}
\examples{

kali()
}
\references{
kali: Tools for habitat and niche modeling (Oliveros-Ramos and Lujan 2016)
}
\author{
Ricardo Oliveros-Ramos Maintainer: Ricardo Oliveros-Ramos
<ricardo.oliveros@gmail.com>
}
\keyword{modeling}
\keyword{niche}
